+++ xcm_integration_tests/emulated/networks/dancelight-system/Cargo.toml
@@ -25,3 +25,4 @@ simple-template-dancelight-emulated-chain = { workspace = true }
 snowbridge-pallet-outbound-queue = { workspace = true }
 tanssi-emulated-integration-tests-common = { workspace = true }
 xcm-emulator = { workspace = true }
+dancelight-emulated-chain = { workspace = true }
