+++ xcm_integration_tests/emulated/tests/assets/dancelight/src/tests/xcm_message_exporter.rs
@@ -15,12 +15,11 @@
 // along with Tanssi.  If not, see <http://www.gnu.org/licenses/>
 
 use {
-    // dancelight_emulated_chain::DancelightRelayPallet,
+    dancelight_emulated_chain::DancelightRelayPallet,
     dancelight_runtime::xcm_config,
     dancelight_runtime_test_utils::ExtBuilder,
     dancelight_runtime_test_utils::ALICE,
     dancelight_system_emulated_network::DancelightRelay as Dancelight,
-    dancelight_system_emulated_network::DancelightRelayPallet,
     frame_support::{assert_err, weights::Weight},
     pallet_xcm::Error,
     primitives::AccountId,
